Function reaper_low::bootstrap_extension_plugin

pub unsafe fn bootstrap_extension_plugin(
    h_instance: HINSTANCE,
    rec: *mut reaper_plugin_info_t,
    static_context: StaticExtensionPluginContext,
    init: fn(_: PluginContext) -> Result<(), Box<dyn Error>>
) -> i32

This is a convenience function for bootstrapping extension plug-ins.

This function basically translates the REAPER extension plug-in main entry point signature (ReaperPluginEntry()) to a typical Rust main entry point signature (main()). It is intended to be used by macros in the reaper-macros crate.

Safety

REAPER can crash if you pass an invalid pointer.