Module reaper_low::raw

Exposes important raw types, functions and constants from the C++ REAPER API.

Structs

ACCEL

AudioAccessor

HDC__

HMENU__

HWND__

IReaperControlSurface

IReaperPitchShift

KbdCmd

KbdSectionInfo

MENUITEMINFO

MIDI_event_t

MIDI event definition and abstract list

MIDI_eventlist

MSG

MediaItem

MediaItem_Take

MediaTrack

PAINTSTRUCT

PCM_sink

PCM_source

PCM_source_peaktransfer_t

PCM_source_transfer_t

PCM source API

POINT

ProjectStateContext

REAPER_Resample_Interface

Resampler API (plug-ins can use this for SRC)

RECT

ReaProject

SCROLLINFO

TrackEnvelope

WDL_HeapBuf

accelerator_register_t

audio_hook_register_t

gaccel_register_t

midi_Input

midi_Output

midi_quantize_mode_t

PCM sink API

midi_realtime_write_struct_t

preview_register_t

reaper_plugin_info_t

Constants

BM_GETCHECK

BM_GETIMAGE

BM_SETCHECK

BM_SETIMAGE

BST_CHECKED

BST_INDETERMINATE

BST_UNCHECKED

CBN_CLOSEUP

CBN_DROPDOWN

CBN_EDITCHANGE

CBN_SELCHANGE

CB_ADDSTRING

CB_DELETESTRING

CB_FINDSTRING

CB_FINDSTRINGEXACT

CB_GETCOUNT

CB_GETCURSEL

CB_GETITEMDATA

CB_GETLBTEXT

CB_GETLBTEXTLEN

CB_INITSTORAGE

CB_INSERTSTRING

CB_RESETCONTENT

CB_SETCURSEL

CB_SETITEMDATA

COLOR_3DDKSHADOW

COLOR_3DFACE

COLOR_3DHILIGHT

COLOR_3DSHADOW

COLOR_BTNFACE

COLOR_BTNTEXT

COLOR_INFOBK

COLOR_INFOTEXT

COLOR_SCROLLBAR

COLOR_WINDOW

CSURF_EXT_RESET

CSURF_EXT_SETBPMANDPLAYRATE

CSURF_EXT_SETFOCUSEDFX

CSURF_EXT_SETFXCHANGE

CSURF_EXT_SETFXENABLED

CSURF_EXT_SETFXOPEN

CSURF_EXT_SETFXPARAM

CSURF_EXT_SETFXPARAM_RECFX

CSURF_EXT_SETINPUTMONITOR

CSURF_EXT_SETLASTTOUCHEDFX

CSURF_EXT_SETPAN_EX

CSURF_EXT_SETPROJECTMARKERCHANGE

CSURF_EXT_SETRECVPAN

CSURF_EXT_SETRECVVOLUME

CSURF_EXT_SETSENDPAN

CSURF_EXT_SETSENDVOLUME

CSURF_EXT_SUPPORTS_EXTENDED_TOUCH

CSURF_EXT_TRACKFX_PRESET_CHANGED

DLL_PROCESS_ATTACH

DLL_PROCESS_DETACH

DT_BOTTOM

DT_CALCRECT

DT_CENTER

DT_END_ELLIPSIS

DT_LEFT

DT_NOCLIP

DT_NOPREFIX

DT_RIGHT

DT_SINGLELINE

DT_TOP

DT_VCENTER

DT_WORDBREAK

EN_CHANGE

EN_KILLFOCUS

EN_SETFOCUS

GMEM_DDESHARE

GMEM_DISCARDABLE

GMEM_FIXED

GMEM_LOWER

GMEM_MOVEABLE

GMEM_SHARE

GMEM_ZEROINIT

IDABORT

IDCANCEL

IDIGNORE

IDNO

IDOK

IDRETRY

IDYES

MB_ICONERROR

MB_ICONINFORMATION

MB_ICONSTOP

MB_OK

MB_OKCANCEL

MB_RETRYCANCEL

MB_YESNO

MB_YESNOCANCEL

MF_BITMAP

MF_BYCOMMAND

MF_BYPOSITION

MF_CHECKED

MF_DISABLED

MF_ENABLED

MF_GRAYED

MF_POPUP

MF_SEPARATOR

MF_STRING

MF_UNCHECKED

MIIM_BITMAP

MIIM_DATA

MIIM_ID

MIIM_STATE

MIIM_SUBMENU

MIIM_TYPE

OPAQUE

PCM_SINK_EXT_CREATESOURCE

PCM_SOURCE_EXT_ADDMIDIEVENTS

PCM_SOURCE_EXT_EXPORTTOFILE

PCM_SOURCE_EXT_GETPOOLEDMIDIID

PCM_SOURCE_EXT_NOTIFYPREVIEWPLAYPOS

PCM_SOURCE_EXT_OPENEDITOR

PCM_SOURCE_EXT_REMOVEFROMMIDIPOOL

PCM_SOURCE_EXT_SETPREVIEWTEMPO

REAPER_PITCHSHIFT_API_VER

REAPER_PLUGIN_VERSION

RESAMPLE_EXT_SETRSMODE

SB_BOTH

SB_BOTTOM

SB_CTL

SB_ENDSCROLL

SB_HORZ

SB_LEFT

SB_LINEDOWN

SB_LINELEFT

SB_LINERIGHT

SB_LINEUP

SB_PAGEDOWN

SB_PAGELEFT

SB_PAGERIGHT

SB_PAGEUP

SB_RIGHT

SB_THUMBPOSITION

SB_THUMBTRACK

SB_TOP

SB_VERT

SIF_ALL

SIF_DISABLENOSCROLL

SIF_PAGE

SIF_POS

SIF_RANGE

SIF_TRACKPOS

SM_CXHSCROLL

SM_CXSCREEN

SM_CXVSCROLL

SM_CYHSCROLL

SM_CYMENU

SM_CYSCREEN

SM_CYVSCROLL

SRCCOPY

SRCCOPY_USEALPHACHAN

SWP_FRAMECHANGED

SWP_NOACTIVATE

SWP_NOCOPYBITS

SWP_NOMOVE

SWP_NOSIZE

SWP_NOZORDER

SWP_SHOWWINDOW

SW_HIDE

SW_NORMAL

SW_RESTORE

SW_SHOW

SW_SHOWDEFAULT

SW_SHOWMAXIMIZED

SW_SHOWMINIMIZED

SW_SHOWNA

SW_SHOWNOACTIVATE

SW_SHOWNORMAL

TBM_GETPOS

TBM_SETPOS

TBM_SETRANGE

TBM_SETSEL

TBM_SETTIC

TPM_BOTTOMALIGN

TPM_CENTERALIGN

TPM_HORIZONTAL

TPM_LEFTALIGN

TPM_LEFTBUTTON

TPM_NONOTIFY

TPM_RETURNCMD

TPM_RIGHTALIGN

TPM_RIGHTBUTTON

TPM_TOPALIGN

TPM_VCENTERALIGN

TPM_VERTICAL

TRANSPARENT

UNDO_STATE_ALL

UNDO_STATE_FREEZE

UNDO_STATE_FX

UNDO_STATE_ITEMS

UNDO_STATE_MISCCFG

UNDO_STATE_TRACKCFG

VK_ADD

VK_BACK

VK_CAPITAL

VK_CLEAR

VK_CONTROL

VK_DECIMAL

VK_DELETE

VK_DIVIDE

VK_DOWN

VK_END

VK_ESCAPE

VK_F1

VK_F2

VK_F3

VK_F4

VK_F5

VK_F6

VK_F7

VK_F8

VK_F9

VK_F10

VK_F11

VK_F12

VK_F13

VK_F14

VK_F15

VK_F16

VK_F17

VK_F18

VK_F19

VK_F20

VK_F21

VK_F22

VK_F23

VK_F24

VK_HELP

VK_HOME

VK_INSERT

VK_LBUTTON

VK_LEFT

VK_LWIN

VK_MBUTTON

VK_MENU

VK_MULTIPLY

VK_NEXT

VK_NUMLOCK

VK_NUMPAD0

VK_NUMPAD1

VK_NUMPAD2

VK_NUMPAD3

VK_NUMPAD4

VK_NUMPAD5

VK_NUMPAD6

VK_NUMPAD7

VK_NUMPAD8

VK_NUMPAD9

VK_PAUSE

VK_PRINT

VK_PRIOR

VK_RBUTTON

VK_RETURN

VK_RIGHT

VK_SCROLL

VK_SELECT

VK_SEPARATOR

VK_SHIFT

VK_SNAPSHOT

VK_SPACE

VK_SUBTRACT

VK_TAB

VK_UP

WM_ACTIVATE

WM_ACTIVATEAPP

WM_CAPTURECHANGED

WM_CHAR

WM_CLOSE

WM_COMMAND

WM_CONTEXTMENU

WM_COPYDATA

WM_CREATE

WM_CTLCOLORBTN

WM_CTLCOLORDLG

WM_CTLCOLOREDIT

WM_CTLCOLORLISTBOX

WM_CTLCOLORMSGBOX

WM_CTLCOLORSCROLLBAR

WM_CTLCOLORSTATIC

WM_DEADCHAR

WM_DESTROY

WM_DISPLAYCHANGE

WM_DRAWITEM

WM_DROPFILES

WM_ERASEBKGND

WM_GESTURE

WM_GETFONT

WM_GETMINMAXINFO

WM_GETOBJECT

WM_HSCROLL

WM_INITDIALOG

WM_INITMENUPOPUP

WM_KEYDOWN

WM_KEYFIRST

WM_KEYLAST

WM_KEYUP

WM_LBUTTONDBLCLK

WM_LBUTTONDOWN

WM_LBUTTONUP

WM_MBUTTONDBLCLK

WM_MBUTTONDOWN

WM_MBUTTONUP

WM_MOUSEACTIVATE

WM_MOUSEFIRST

WM_MOUSEHWHEEL

WM_MOUSELAST

WM_MOUSEMOVE

WM_MOUSEWHEEL

WM_MOVE

WM_NCCALCSIZE

WM_NCDESTROY

WM_NCHITTEST

WM_NCLBUTTONDBLCLK

WM_NCLBUTTONDOWN

WM_NCLBUTTONUP

WM_NCMBUTTONDBLCLK

WM_NCMBUTTONDOWN

WM_NCMBUTTONUP

WM_NCMOUSEMOVE

WM_NCPAINT

WM_NCRBUTTONDBLCLK

WM_NCRBUTTONDOWN

WM_NCRBUTTONUP

WM_NOTIFY

WM_PAINT

WM_RBUTTONDBLCLK

WM_RBUTTONDOWN

WM_RBUTTONUP

WM_SETCURSOR

WM_SETFONT

WM_SETREDRAW

WM_SETTEXT

WM_SHOWWINDOW

WM_SIZE

WM_STYLECHANGED

WM_SYSCHAR

WM_SYSCOMMAND

WM_SYSDEADCHAR

WM_SYSKEYDOWN

WM_SYSKEYUP

WM_TIMER

WM_USER

WM_VSCROLL

Type Definitions

ApiVararg

Function pointer type for exposing custom API functions to ReaScript.

GUID

HANDLE

HBRUSH

HDC

HINSTANCE

HMENU

HWND

HookCommand

Function pointer type for hook commands.

HookCommand2

Function pointer type for hook commands with MIDI/OSC support.

HookPostCommand

Function pointer type for getting notified about invocation of hook command.

HookPostCommand2

Function pointer type for getting notified about invocation of hook command 2.

INT_PTR

LPARAM

LPSTR

LRESULT

ReaSample

TimerFunction

Function pointer type for timers.

ToggleAction

Function pointer type for toggle actions.

UINT

ULONG_PTR

WDL_INT64

WPARAM

Unions

pthread_mutex_t