Struct reaper_low::raw::PCM_source

#[repr(C)]
pub struct PCM_source {
    pub vtable_: *const PCM_source__bindgen_vtable,
}

Fields
vtable_: *const PCM_source__bindgen_vtable
Implementations

impl PCM_source

pub fn GetLength(&self) -> f64

pub fn IsAvailable(&self) -> bool

pub fn Duplicate(&self) -> *mut PCM_source

pub fn GetType(&self) -> *const c_char

pub fn GetFileName(&self) -> *const c_char

pub fn GetSource(&self) -> *mut PCM_source

pub unsafe fn Extended(
&self,
call: c_int,
parm1: *mut c_void,
parm2: *mut c_void,
parm3: *mut c_void
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SetAvailable(&self, avail: bool)

pub unsafe fn SetFileName(&self, newfn: *const c_char) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetSource(&self, src: *mut PCM_source)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetNumChannels(&self) -> c_int

pub fn GetSampleRate(&self) -> f64

pub fn GetLengthBeats(&self) -> f64

pub fn GetBitsPerSample(&self) -> c_int

pub fn GetPreferredPosition(&self) -> f64

pub unsafe fn PropertiesWindow(&self, hwndParent: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetSamples(&self, block: *mut PCM_source_transfer_t)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetPeakInfo(&self, block: *mut PCM_source_peaktransfer_t)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SaveState(&self, ctx: *mut ProjectStateContext)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn LoadState(
&self,
firstline: *const c_char,
ctx: *mut ProjectStateContext
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub fn Peaks_Clear(&self, deleteFile: bool)

pub fn PeaksBuild_Begin(&self) -> c_int

pub fn PeaksBuild_Run(&self) -> c_int

pub fn PeaksBuild_Finish(&self)