Struct reaper_low::PluginContext

pub struct PluginContext { /* private fields */ }

This represents the context which is needed to access REAPER functions from plug-ins.

Once obtained, it is supposed to be passed to Reaper::load().

Implementations

impl PluginContext

pub unsafe fn from_extension_plugin(
h_instance: HINSTANCE,
rec: reaper_plugin_info_t,
static_context: StaticExtensionPluginContext
) -> Result<PluginContext, ContextFromExtensionPluginError>

Creates a plug-in context from an extension entry point plug-in info.

It requires a module handle and the pointer to a reaper_plugin_info_t struct. REAPER provides both when calling the ReaperPluginEntry function (the main entry point for any extension plug-in).

It’s recommended to use the reaper_extension_plugin macro in the reaper-macros crate instead of calling this function directly.

Errors

Returns an error if the given plug-in info is not suitable for loading REAPER functions.

Safety

REAPER can crash if you pass an invalid pointer.

pub fn from_vst_plugin(
host: &HostCallback,
static_context: StaticVstPluginContext
) -> Result<PluginContext, ContextFromVstPluginError>

Creates a plug-in context from a VST host callback.

It requires the host callback which vst-rs passes to the plugin’s new() function.

Errors

Returns an error if the given host callback is not suitable for loading REAPER functions.

pub unsafe fn GetFunc(&self, name: *const c_char) -> *mut c_void

Returns a generic API function by its name.

Safety

REAPER can crash if you pass an invalid pointer.

pub fn h_instance(&self) -> HINSTANCE

On Windows, this returns the HINSTANCE passed to DllMain (VST plug-ins) or ReaperPluginEntry (extension plug-ins).

The returned HINSTANCE represents the handle of the module (DLL) containing the plug-in.

On Linux, this returns null.

pub fn swell_function_provider(
&self
) -> Option<unsafe extern "C" fn(name: *const c_char) -> *mut c_void>

On Linux, this returns a pointer to a function for getting a generic SWELL API function by its name.

On Windows, this returns None.

pub fn type_specific(&self) -> &TypeSpecificPluginContext

Returns the type-specific plug-in context.

pub fn is_in_main_thread(&self) -> bool

Returns whether we are currently in the main thread.