Struct reaper_low::Swell

pub struct Swell { /* private fields */ }

This is the low-level API access point to all SWELL functions.

SWELL is the Simple Windows Emulation Layer and is exposed by REAPER for Linux and Mac OS X.

See Reaper for details how to use this struct (it’s very similar).

Implementations

impl Swell

pub fn load(plugin_context: PluginContext) -> Swell

Loads all available SWELL functions from the given plug-in context.

Returns a Swell instance which allows you to call these functions.

On Windows, this function will not load any function pointers because the methods in this struct delegate to the corresponding Windows functions.

Panics

If this is Linux and the SWELL function provider is not available, this function panics.

pub unsafe fn lstrcpyn(
&self,
dest: *mut c_char,
src: *const c_char,
l: c_int
) -> *mut c_char

Safety

REAPER can crash if you pass an invalid pointer.

pub fn MulDiv(&self, arg1: c_int, arg2: c_int, arg3: c_int) -> c_int

pub fn Sleep(&self, ms: c_int)

pub fn GetTickCount(&self) -> c_uint

pub unsafe fn GetFileTime(
&self,
filedes: c_int,
lpCreationTime: *mut FILETIME,
lpLastAccessTime: *mut FILETIME,
lpLastWriteTime: *mut FILETIME
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WritePrivateProfileString(
&self,
appname: *const c_char,
keyname: *const c_char,
val: *const c_char,
fn_: *const c_char
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetPrivateProfileString(
&self,
appname: *const c_char,
keyname: *const c_char,
def: *const c_char,
ret: *mut c_char,
retsize: c_int,
fn_: *const c_char
) -> c_uint

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetPrivateProfileInt(
&self,
appname: *const c_char,
keyname: *const c_char,
def: c_int,
fn_: *const c_char
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetPrivateProfileStruct(
&self,
appname: *const c_char,
keyname: *const c_char,
buf: *mut c_void,
bufsz: c_int,
fn_: *const c_char
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WritePrivateProfileStruct(
&self,
appname: *const c_char,
keyname: *const c_char,
buf: *const c_void,
bufsz: c_int,
fn_: *const c_char
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WritePrivateProfileSection(
&self,
appname: *const c_char,
strings: *const c_char,
fn_: *const c_char
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetPrivateProfileSection(
&self,
appname: *const c_char,
strout: *mut c_char,
strout_len: c_uint,
fn_: *const c_char
) -> c_uint

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetModuleFileName(
&self,
hInst: HINSTANCE,
fn_: *mut c_char,
nSize: c_uint
) -> c_uint

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_PtInRect(&self, r: *const RECT, p: POINT) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ShellExecute(
&self,
hwndDlg: HWND,
action: *const c_char,
content1: *const c_char,
content2: *const c_char,
content3: *const c_char,
blah: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn MessageBox(
&self,
hwndParent: HWND,
text: *const c_char,
caption: *const c_char,
type_: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn BrowseForFiles(
&self,
text: *const c_char,
initialdir: *const c_char,
initialfile: *const c_char,
allowmul: bool,
extlist: *const c_char
) -> *mut c_char

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn BrowseForSaveFile(
&self,
text: *const c_char,
initialdir: *const c_char,
initialfile: *const c_char,
extlist: *const c_char,
fn_: *mut c_char,
fnsize: c_int
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn BrowseForDirectory(
&self,
text: *const c_char,
initialdir: *const c_char,
fn_: *mut c_char,
fnsize: c_int
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn BrowseFile_SetTemplate(
&self,
dlgid: *const c_char,
dlgProc: Option<unsafe extern "C" fn(arg1: HWND, arg2: UINT, arg3: WPARAM, arg4: LPARAM) -> INT_PTR>,
reshead: *mut SWELL_DialogResourceIndex
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetDlgItem(&self, arg1: HWND, arg2: c_int) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ShowWindow(&self, arg1: HWND, arg2: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DestroyWindow(&self, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetGestureInfo(
&self,
lParam: LPARAM,
gi: *mut tagGESTUREINFO
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_HideApp(&self)

pub unsafe fn SetDlgItemText(
&self,
arg1: HWND,
idx: c_int,
text: *const c_char
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetDlgItemInt(
&self,
arg1: HWND,
idx: c_int,
val: c_int,
issigned: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetDlgItemInt(
&self,
arg1: HWND,
idx: c_int,
translated: *mut c_schar,
issigned: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetDlgItemText(
&self,
arg1: HWND,
idx: c_int,
text: *mut c_char,
textlen: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindowTextLength(&self, arg1: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CheckDlgButton(&self, hwnd: HWND, idx: c_int, check: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn IsDlgButtonChecked(&self, hwnd: HWND, idx: c_int) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnableWindow(&self, hwnd: HWND, enable: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetFocus(&self, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetFocus(&self) -> HWND

pub unsafe fn SetForegroundWindow(&self, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetForegroundWindow(&self) -> HWND

pub unsafe fn SetCapture(&self, hwnd: HWND) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetCapture(&self) -> HWND

pub fn ReleaseCapture(&self)

pub unsafe fn IsChild(&self, hwndParent: HWND, hwndChild: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetParent(&self, hwnd: HWND) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetParent(&self, hwnd: HWND, newPar: HWND) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindow(&self, hwnd: HWND, what: c_int) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnumWindows(
&self,
proc_: Option<unsafe extern "C" fn(arg1: HWND, arg2: LPARAM) -> c_schar>,
lp: LPARAM
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn FindWindowEx(
&self,
par: HWND,
lastw: HWND,
classname: *const c_char,
title: *const c_char
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ClientToScreen(&self, hwnd: HWND, p: *mut POINT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ScreenToClient(&self, hwnd: HWND, p: *mut POINT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindowRect(&self, hwnd: HWND, r: *mut RECT) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindowContentViewRect(&self, hwnd: HWND, r: *mut RECT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetClientRect(&self, hwnd: HWND, r: *mut RECT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WindowFromPoint(&self, p: POINT) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WinOffsetRect(
&self,
lprc: *mut RECT,
dx: c_int,
dy: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WinSetRect(
&self,
lprc: *mut RECT,
xLeft: c_int,
yTop: c_int,
xRight: c_int,
yBottom: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WinUnionRect(
&self,
out: *mut RECT,
in1: *const RECT,
in2: *const RECT
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WinIntersectRect(
&self,
out: *mut RECT,
in1: *const RECT,
in2: *const RECT
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetWindowPos(
&self,
hwnd: HWND,
unused: HWND,
x: c_int,
y: c_int,
cx: c_int,
cy: c_int,
flags: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetWindowLevel(&self, hwnd: HWND, newlevel: c_int) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn InvalidateRect(
&self,
hwnd: HWND,
r: *const RECT,
eraseBk: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn UpdateWindow(&self, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindowLong(&self, hwnd: HWND, idx: c_int) -> isize

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetWindowLong(&self, hwnd: HWND, idx: c_int, val: isize) -> isize

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ScrollWindow(
&self,
hwnd: HWND,
xamt: c_int,
yamt: c_int,
lpRect: *const RECT,
lpClipRect: *const RECT
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnumPropsEx(
&self,
arg1: HWND,
arg2: Option<unsafe extern "C" fn(hwnd: HWND, lpszString: *const c_char, hData: HANDLE, lParam: LPARAM) -> c_schar>,
arg3: LPARAM
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetProp(&self, arg1: HWND, arg2: *const c_char) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetProp(
&self,
arg1: HWND,
arg2: *const c_char,
arg3: HANDLE
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn RemoveProp(&self, arg1: HWND, arg2: *const c_char) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn IsWindowVisible(&self, hwnd: HWND) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn IsWindow(&self, hwnd: HWND) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetTimer(
&self,
hwnd: HWND,
timerid: usize,
rate: UINT,
tProc: Option<unsafe extern "C" fn(hwnd: HWND, uMsg: UINT, idEvent: usize, dwTime: c_uint)>
) -> usize

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn KillTimer(&self, hwnd: HWND, timerid: usize) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetExtendedListViewStyleEx(
&self,
h: HWND,
mask: c_int,
style: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_InsertColumn(
&self,
h: HWND,
pos: c_int,
lvc: *const LVCOLUMN
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_DeleteColumn(&self, h: HWND, pos: c_int) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetColumn(&self, h: HWND, pos: c_int, lvc: *const LVCOLUMN)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetColumnWidth(&self, h: HWND, pos: c_int) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_InsertItem(&self, h: HWND, item: *const LVITEM) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetItemText(
&self,
h: HWND,
ipos: c_int,
cpos: c_int,
txt: *const c_char
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetItem(&self, h: HWND, item: *mut LVITEM) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetNextItem(
&self,
h: HWND,
istart: c_int,
flags: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetItem(&self, h: HWND, item: *mut LVITEM) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetItemState(
&self,
h: HWND,
ipos: c_int,
mask: UINT
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_DeleteItem(&self, h: HWND, ipos: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_DeleteAllItems(&self, h: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetSelectedCount(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetItemCount(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetSelectionMark(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetColumnWidth(&self, h: HWND, colpos: c_int, wid: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetItemState(
&self,
h: HWND,
item: c_int,
state: UINT,
statemask: UINT
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_RedrawItems(
&self,
h: HWND,
startitem: c_int,
enditem: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetItemCount(&self, h: HWND, cnt: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_EnsureVisible(&self, h: HWND, i: c_int, pok: c_schar)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetImageList(
&self,
h: HWND,
imagelist: *mut HIMAGELIST__,
which: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SubItemHitTest(
&self,
h: HWND,
pinf: *mut LVHITTESTINFO
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetItemText(
&self,
hwnd: HWND,
item: c_int,
subitem: c_int,
text: *mut c_char,
textmax: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SortItems(
&self,
hwnd: HWND,
compf: Option<unsafe extern "C" fn(arg1: LPARAM, arg2: LPARAM, arg3: LPARAM) -> c_int>,
parm: LPARAM
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_Scroll(
&self,
h: HWND,
xscroll: c_int,
yscroll: c_int
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetTopIndex(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetCountPerPage(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetColumnOrderArray(
&self,
h: HWND,
cnt: c_int,
arr: *mut c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetColumnOrderArray(
&self,
h: HWND,
cnt: c_int,
arr: *mut c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetHeader(&self, h: HWND) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn Header_GetItemCount(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn Header_GetItem(
&self,
h: HWND,
col: c_int,
hi: *mut HDITEM
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn Header_SetItem(
&self,
h: HWND,
col: c_int,
hi: *mut HDITEM
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetItemRect(
&self,
h: HWND,
item: c_int,
r: *mut RECT,
code: c_int
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_GetSubItemRect(
&self,
h: HWND,
item: c_int,
subitem: c_int,
code: c_int,
r: *mut RECT
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_HitTest(&self, h: HWND, pinf: *mut LVHITTESTINFO) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetListViewHeaderHeight(&self, h: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub fn ImageList_CreateEx(&self) -> *mut HIMAGELIST__

pub unsafe fn ImageList_Remove(
&self,
list: *mut HIMAGELIST__,
idx: c_int
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ImageList_ReplaceIcon(
&self,
list: *mut HIMAGELIST__,
offset: c_int,
image: *mut HGDIOBJ__
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ImageList_Add(
&self,
list: *mut HIMAGELIST__,
image: *mut HGDIOBJ__,
mask: *mut HGDIOBJ__
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ImageList_Destroy(&self, arg1: *mut HIMAGELIST__)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TabCtrl_GetItemCount(&self, hwnd: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TabCtrl_DeleteItem(&self, hwnd: HWND, idx: c_int) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TabCtrl_InsertItem(
&self,
hwnd: HWND,
idx: c_int,
item: *mut TCITEM
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TabCtrl_SetCurSel(&self, hwnd: HWND, idx: c_int) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TabCtrl_GetCurSel(&self, hwnd: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TabCtrl_AdjustRect(
&self,
hwnd: HWND,
fLarger: c_schar,
r: *mut RECT
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_InsertItem(
&self,
hwnd: HWND,
ins: *mut TVINSERTSTRUCT
) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_Expand(
&self,
hwnd: HWND,
item: *mut HTREEITEM__,
flag: UINT
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_GetSelection(&self, hwnd: HWND) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_DeleteItem(&self, hwnd: HWND, item: *mut HTREEITEM__)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_DeleteAllItems(&self, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_SelectItem(&self, hwnd: HWND, item: *mut HTREEITEM__)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_EnsureVisible(&self, hwnd: HWND, item: *mut HTREEITEM__)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_GetItem(&self, hwnd: HWND, pitem: *mut TVITEM) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_SetItem(&self, hwnd: HWND, pitem: *mut TVITEM) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_HitTest(
&self,
hwnd: HWND,
hti: *mut TVHITTESTINFO
) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_SetIndent(&self, hwnd: HWND, indent: c_int) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_GetParent(
&self,
hwnd: HWND,
item: *mut HTREEITEM__
) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_GetChild(
&self,
hwnd: HWND,
item: *mut HTREEITEM__
) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_GetNextSibling(
&self,
hwnd: HWND,
item: *mut HTREEITEM__
) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_GetRoot(&self, hwnd: HWND) -> *mut HTREEITEM__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_SetBkColor(&self, hwnd: HWND, color: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TreeView_SetTextColor(&self, hwnd: HWND, color: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetBkColor(&self, hwnd: HWND, color: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetTextBkColor(&self, hwnd: HWND, color: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetTextColor(&self, hwnd: HWND, color: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetGridColor(&self, hwnd: HWND, color: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ListView_SetSelColors(
&self,
hwnd: HWND,
colors: *mut c_int,
ncolors: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_ModalWindowStart(&self, hwnd: HWND) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_ModalWindowRun(
&self,
ctx: *mut c_void,
ret: *mut c_int
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_ModalWindowEnd(&self, ctx: *mut c_void)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_CloseWindow(&self, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn CreatePopupMenu(&self) -> HMENU

pub unsafe fn CreatePopupMenuEx(&self, title: *const c_char) -> HMENU

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DestroyMenu(&self, hMenu: HMENU)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn AddMenuItem(
&self,
hMenu: HMENU,
pos: c_int,
name: *const c_char,
tagid: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetSubMenu(&self, hMenu: HMENU, pos: c_int) -> HMENU

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetMenuItemCount(&self, hMenu: HMENU) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetMenuItemID(&self, hMenu: HMENU, pos: c_int) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetMenuItemModifier(
&self,
hMenu: HMENU,
idx: c_int,
flag: c_int,
code: c_int,
mask: c_uint
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetMenuItemText(
&self,
hMenu: HMENU,
idx: c_int,
flag: c_int,
text: *const c_char
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnableMenuItem(&self, hMenu: HMENU, idx: c_int, en: c_int) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DeleteMenu(&self, hMenu: HMENU, idx: c_int, flag: c_int) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CheckMenuItem(&self, hMenu: HMENU, idx: c_int, chk: c_int) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn InsertMenuItem(
&self,
hMenu: HMENU,
pos: c_int,
byPos: c_schar,
mi: *mut MENUITEMINFO
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_InsertMenu(
&self,
menu: HMENU,
pos: c_int,
flag: c_uint,
idx: usize,
str: *const c_char
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetMenuItemInfo(
&self,
hMenu: HMENU,
pos: c_int,
byPos: c_schar,
mi: *mut MENUITEMINFO
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetMenuItemInfo(
&self,
hMenu: HMENU,
pos: c_int,
byPos: c_schar,
mi: *mut MENUITEMINFO
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DrawMenuBar(&self, arg1: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_LoadMenu(
&self,
head: *mut SWELL_MenuResourceIndex,
resid: *const c_char
) -> HMENU

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn TrackPopupMenu(
&self,
hMenu: HMENU,
flags: c_int,
xpos: c_int,
ypos: c_int,
resvd: c_int,
hwnd: HWND,
r: *const RECT
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetMenuDestination(&self, menu: HMENU, hwnd: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_DuplicateMenu(&self, menu: HMENU) -> HMENU

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetMenu(&self, hwnd: HWND, menu: HMENU) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetMenu(&self, hwnd: HWND) -> HMENU

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_GetDefaultWindowMenu(&self) -> HMENU

pub unsafe fn SWELL_SetDefaultWindowMenu(&self, arg1: HMENU)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_GetDefaultModalWindowMenu(&self) -> HMENU

pub unsafe fn SWELL_SetDefaultModalWindowMenu(&self, arg1: HMENU)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_GetCurrentMenu(&self) -> HMENU

pub unsafe fn SWELL_SetCurrentMenu(&self, arg1: HMENU)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_DialogBox(
&self,
reshead: *mut SWELL_DialogResourceIndex,
resid: *const c_char,
parent: HWND,
dlgproc: Option<unsafe extern "C" fn(arg1: HWND, arg2: UINT, arg3: WPARAM, arg4: LPARAM) -> INT_PTR>,
param: LPARAM
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_CreateDialog(
&self,
reshead: *mut SWELL_DialogResourceIndex,
resid: *const c_char,
parent: HWND,
dlgproc: Option<unsafe extern "C" fn(arg1: HWND, arg2: UINT, arg3: WPARAM, arg4: LPARAM) -> INT_PTR>,
param: LPARAM
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_RegisterCustomControlCreator(
&self,
proc_: Option<unsafe extern "C" fn(parent: HWND, cname: *const c_char, idx: c_int, classname: *const c_char, style: c_int, x: c_int, y: c_int, w: c_int, h: c_int) -> HWND>
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_UnregisterCustomControlCreator(
&self,
proc_: Option<unsafe extern "C" fn(parent: HWND, cname: *const c_char, idx: c_int, classname: *const c_char, style: c_int, x: c_int, y: c_int, w: c_int, h: c_int) -> HWND>
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DefWindowProc(
&self,
hwnd: HWND,
msg: UINT,
wParam: WPARAM,
lParam: LPARAM
) -> LRESULT

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EndDialog(&self, arg1: HWND, arg2: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetDefaultButtonID(
&self,
hwndDlg: HWND,
onlyIfEnabled: bool
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SendMessage(
&self,
arg1: HWND,
arg2: UINT,
arg3: WPARAM,
arg4: LPARAM
) -> LRESULT

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_BroadcastMessage(&self, arg1: UINT, arg2: WPARAM, arg3: LPARAM)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn PostMessage(
&self,
hwnd: HWND,
msg: UINT,
wParam: WPARAM,
lParam: LPARAM
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_MessageQueue_Flush(&self)

pub unsafe fn SWELL_MessageQueue_Clear(&self, h: HWND)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_KeyToASCII(
&self,
wParam: c_int,
lParam: c_int,
newflags: *mut c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetAsyncKeyState(&self, key: c_int) -> c_ushort

pub unsafe fn GetCursorPos(&self, pt: *mut POINT)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetMessagePos(&self) -> c_uint

pub unsafe fn SWELL_LoadCursor(&self, idx: *const c_char) -> *mut HCURSOR__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetCursor(&self, curs: *mut HCURSOR__)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_EnableRightClickEmulate(&self, enable: c_schar)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_GetCursor(&self) -> *mut HCURSOR__

pub fn SWELL_GetLastSetCursor(&self) -> *mut HCURSOR__

pub fn SWELL_IsCursorVisible(&self) -> bool

pub unsafe fn SWELL_ShowCursor(&self, bShow: c_schar) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_SetCursorPos(&self, X: c_int, Y: c_int) -> c_schar

pub unsafe fn SWELL_GetViewPort(
&self,
r: *mut RECT,
sourcerect: *const RECT,
wantWork: bool
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn OpenClipboard(&self, hwndDlg: HWND) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub fn CloseClipboard(&self)

pub unsafe fn GetClipboardData(&self, type_: UINT) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub fn EmptyClipboard(&self)

pub unsafe fn SetClipboardData(&self, type_: UINT, h: HANDLE)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn RegisterClipboardFormat(&self, desc: *const c_char) -> UINT

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnumClipboardFormats(&self, lastfmt: UINT) -> UINT

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GlobalAlloc(&self, flags: c_int, sz: c_int) -> HANDLE

pub unsafe fn GlobalLock(&self, h: HANDLE) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GlobalSize(&self, h: HANDLE) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GlobalUnlock(&self, h: HANDLE)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GlobalFree(&self, h: HANDLE)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CreateThread(
&self,
TA: *mut c_void,
stackSize: c_uint,
ThreadProc: Option<unsafe extern "C" fn(arg1: *mut c_void) -> c_uint>,
parm: *mut c_void,
cf: c_uint,
tidOut: *mut c_uint
) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CreateEvent(
&self,
SA: *mut c_void,
manualReset: c_schar,
initialSig: c_schar,
ignored: *const c_char
) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CreateEventAsSocket(
&self,
SA: *mut c_void,
manualReset: c_schar,
initialSig: c_schar,
ignored: *const c_char
) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetCurrentThreadId(&self) -> c_uint

pub unsafe fn WaitForSingleObject(&self, hand: HANDLE, msTO: c_uint) -> c_uint

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn WaitForAnySocketObject(
&self,
numObjs: c_int,
objs: *mut HANDLE,
msTO: c_uint
) -> c_uint

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CloseHandle(&self, hand: HANDLE) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetThreadPriority(&self, evt: HANDLE, prio: c_int) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetEvent(&self, evt: HANDLE) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ResetEvent(&self, evt: HANDLE) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_CreateProcessFromPID(&self, pid: c_int) -> HANDLE

pub unsafe fn SWELL_CreateProcess(
&self,
exe: *const c_char,
nparams: c_int,
params: *mut *const c_char
) -> HANDLE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetProcessExitCode(&self, hand: HANDLE) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn LoadLibraryGlobals(
&self,
fileName: *const c_char,
symbolsAsGlobals: bool
) -> HINSTANCE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn LoadLibrary(&self, fileName: *const c_char) -> HINSTANCE

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetProcAddress(
&self,
hInst: HINSTANCE,
procName: *const c_char
) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn FreeLibrary(&self, hInst: HINSTANCE) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetBundle(&self, hInst: HINSTANCE) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_CreateMemContext(&self, hdc: HDC, w: c_int, h: c_int) -> HDC

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_DeleteGfxContext(&self, arg1: HDC)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetCtxGC(&self, ctx: HDC) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetCtxFrameBuffer(&self, ctx: HDC) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_PushClipRegion(&self, ctx: HDC)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetClipRegion(&self, ctx: HDC, r: *const RECT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_PopClipRegion(&self, ctx: HDC)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CreateFontIndirect(&self, arg1: *mut LOGFONT) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CreateFont(
&self,
lfHeight: c_int,
lfWidth: c_int,
lfEscapement: c_int,
lfOrientation: c_int,
lfWeight: c_int,
lfItalic: c_char,
lfUnderline: c_char,
lfStrikeOut: c_char,
lfCharSet: c_char,
lfOutPrecision: c_char,
lfClipPrecision: c_char,
lfQuality: c_char,
lfPitchAndFamily: c_char,
lfFaceName: *const c_char
) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub fn CreatePen(&self, attr: c_int, wid: c_int, col: c_int) -> *mut HGDIOBJ__

pub fn CreateSolidBrush(&self, col: c_int) -> HBRUSH

pub fn CreatePenAlpha(
&self,
attr: c_int,
wid: c_int,
col: c_int,
alpha: f32
) -> *mut HGDIOBJ__

pub fn CreateSolidBrushAlpha(&self, col: c_int, alpha: f32) -> HBRUSH

pub unsafe fn SelectObject(&self, ctx: HDC, pen: *mut HGDIOBJ__) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetStockObject(&self, wh: c_int) -> *mut HGDIOBJ__

pub unsafe fn DeleteObject(&self, arg1: *mut HGDIOBJ__)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_FillRect(&self, ctx: HDC, r: *const RECT, br: HBRUSH)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn Rectangle(&self, ctx: HDC, l: c_int, t: c_int, r: c_int, b: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn Ellipse(&self, ctx: HDC, l: c_int, t: c_int, r: c_int, b: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_Polygon(&self, ctx: HDC, pts: *mut POINT, npts: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn MoveToEx(&self, ctx: HDC, x: c_int, y: c_int, op: *mut POINT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_LineTo(&self, ctx: HDC, x: c_int, y: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetPixel(&self, ctx: HDC, x: c_int, y: c_int, c: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn PolyBezierTo(&self, ctx: HDC, pts: *mut POINT, np: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_DrawText(
&self,
ctx: HDC,
buf: *const c_char,
len: c_int,
r: *mut RECT,
align: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetTextColor(&self, ctx: HDC, col: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetTextColor(&self, ctx: HDC) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetBkColor(&self, ctx: HDC, col: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetBkMode(&self, ctx: HDC, col: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetGlyphIndicesW(
&self,
ctx: HDC,
buf: *mut u32,
len: c_int,
indices: *mut c_ushort,
flags: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn RoundRect(
&self,
ctx: HDC,
x: c_int,
y: c_int,
x2: c_int,
y2: c_int,
xrnd: c_int,
yrnd: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn PolyPolyline(
&self,
ctx: HDC,
pts: *mut POINT,
cnts: *mut c_uint,
nseg: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetTextMetrics(&self, ctx: HDC, tm: *mut TEXTMETRIC) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetTextFace(
&self,
ctx: HDC,
nCount: c_int,
lpFaceName: *mut c_char
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetObject(
&self,
icon: *mut HGDIOBJ__,
bmsz: c_int,
_bm: *mut c_void
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn CreateIconIndirect(
&self,
iconinfo: *mut _ICONINFO
) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn LoadNamedImage(
&self,
name: *const c_char,
alphaFromMask: bool
) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DrawImageInRect(&self, ctx: HDC, img: *mut HGDIOBJ__, r: *const RECT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn BitBlt(
&self,
hdcOut: HDC,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
hdcIn: HDC,
xin: c_int,
yin: c_int,
mode: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn StretchBlt(
&self,
hdcOut: HDC,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
hdcIn: HDC,
xin: c_int,
yin: c_int,
srcw: c_int,
srch: c_int,
mode: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn StretchBltFromMem(
&self,
hdcOut: HDC,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
bits: *const c_void,
srcw: c_int,
srch: c_int,
srcspan: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_GetScaling256(&self) -> c_int

pub unsafe fn SWELL_ExtendedAPI(
&self,
key: *const c_char,
v: *mut c_void
) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetSysColor(&self, idx: c_int) -> c_int

pub unsafe fn CreateBitmap(
&self,
width: c_int,
height: c_int,
numplanes: c_int,
bitsperpixel: c_int,
bits: *mut c_uchar
) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetOpaque(&self, h: HWND, isopaque: bool)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetAllowNoMiddleManRendering(&self, h: HWND, allow: bool)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetViewGL(&self, h: HWND, wantGL: c_char)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetViewGL(&self, h: HWND) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetGLContextToView(&self, h: HWND) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn BeginPaint(&self, arg1: HWND, arg2: *mut PAINTSTRUCT) -> HDC

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EndPaint(&self, arg1: HWND, arg2: *mut PAINTSTRUCT) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetDC(&self, arg1: HWND) -> HDC

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindowDC(&self, arg1: HWND) -> HDC

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn ReleaseDC(&self, arg1: HWND, arg2: HDC)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_FillDialogBackground(
&self,
hdc: HDC,
r: *const RECT,
level: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_CloneGDIObject(&self, a: *mut HGDIOBJ__) -> *mut HGDIOBJ__

Safety

REAPER can crash if you pass an invalid pointer.

pub fn GetSystemMetrics(&self, arg1: c_int) -> c_int

pub unsafe fn DragQueryPoint(
&self,
arg1: *mut c_void,
arg2: *mut POINT
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DragFinish(&self, arg1: *mut c_void)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DragQueryFile(
&self,
arg1: *mut c_void,
arg2: UINT,
arg3: *mut c_char,
arg4: UINT
) -> UINT

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_InitiateDragDrop(
&self,
arg1: HWND,
srcrect: *mut RECT,
srcfn: *const c_char,
callback: Option<unsafe extern "C" fn(droppath: *const c_char)>
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_InitiateDragDropOfFileList(
&self,
arg1: HWND,
srcrect: *mut RECT,
srclist: *mut *const c_char,
srccount: c_int,
icon: *mut HGDIOBJ__
)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_FinishDragDrop(&self)

pub unsafe fn SWELL_DrawFocusRect(
&self,
hwndPar: HWND,
rct: *mut RECT,
handle: *mut *mut c_void
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_MakeSetCurParms(
&self,
xscale: f32,
yscale: f32,
xtrans: f32,
ytrans: f32,
parent: HWND,
doauto: bool,
dosizetofit: bool
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_MakeButton(
&self,
def: c_int,
label: *const c_char,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
flags: c_int
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_MakeEditField(
&self,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
flags: c_int
) -> HWND

pub unsafe fn SWELL_MakeLabel(
&self,
align: c_int,
label: *const c_char,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
flags: c_int
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_MakeControl(
&self,
cname: *const c_char,
idx: c_int,
classname: *const c_char,
style: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
exstyle: c_int
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_MakeCombo(
&self,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
flags: c_int
) -> HWND

pub unsafe fn SWELL_MakeGroupBox(
&self,
name: *const c_char,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
style: c_int
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_MakeCheckBox(
&self,
name: *const c_char,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
flags: c_int
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_MakeListBox(
&self,
idx: c_int,
x: c_int,
y: c_int,
w: c_int,
h: c_int,
styles: c_int
) -> HWND

pub unsafe fn SWELL_Menu_AddMenuItem(
&self,
hMenu: HMENU,
name: *const c_char,
idx: c_int,
flags: c_uint
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GenerateMenuFromList(
&self,
hMenu: HMENU,
list: *const c_void,
listsz: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GenerateDialogFromList(
&self,
list: *const c_void,
listsz: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn _controlfp(&self, flag: c_uint, mask: c_uint) -> c_uint

pub fn SWELL_Internal_PostMessage_Init(&self)

pub unsafe fn SWELL_LoadCursorFromFile(
&self,
fn_: *const c_char
) -> *mut HCURSOR__

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetWindowWantRaiseAmt(&self, h: HWND, amt: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetWindowWantRaiseAmt(&self, arg1: HWND) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetListViewFastClickMask(&self, hList: HWND, mask: c_int)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetTempPath(&self, sz: c_int, buf: *mut c_char)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_initargs(&self, argc: *mut c_int, argv: *mut *mut *mut c_char)

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_RunMessageLoop(&self)

pub unsafe fn SWELL_CreateXBridgeWindow(
&self,
viewpar: HWND,
wref: *mut *mut c_void,
arg1: *const RECT
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetOSWindow(
&self,
hwnd: HWND,
type_: *const c_char
) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetOSEvent(&self, type_: *const c_char) -> *mut c_void

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GenerateGUID(&self, g: *mut c_void) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnumChildWindows(
&self,
hwnd: HWND,
cwEnumFunc: Option<unsafe extern "C" fn(arg1: HWND, arg2: LPARAM) -> c_schar>,
lParam: LPARAM
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_IsGroupBox(&self, arg1: HWND) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_IsButton(&self, arg1: HWND) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_IsStaticText(&self, arg1: HWND) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_GetDesiredControlSize(&self, hwnd: HWND, r: *mut RECT)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn AddFontResourceEx(
&self,
str: *const c_char,
fl: c_uint,
pdv: *mut c_void
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_Register_Cursor_Resource(
&self,
idx: *const c_char,
name: *const c_char,
hotspot_x: c_int,
hotspot_y: c_int
)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_ChooseColor(
&self,
arg1: HWND,
arg2: *mut c_uint,
ncustom: c_int,
custom: *mut c_uint
) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_ChooseFont(&self, arg1: HWND, arg2: *mut LOGFONT) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn IsWindowEnabled(&self, arg1: HWND) -> bool

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetClassName(
&self,
arg1: HWND,
arg2: *mut c_char,
arg3: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_SetClassName(&self, arg1: HWND, arg2: *const c_char)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SWELL_DisableContextMenu(&self, arg1: HWND, arg2: bool)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn EnumDisplayMonitors(
&self,
arg1: HDC,
arg2: *mut RECT,
arg3: Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: HDC, arg3: *mut RECT, arg4: LPARAM) -> c_schar>,
arg4: LPARAM
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetMonitorInfo(
&self,
arg1: *mut c_void,
arg2: *mut c_void
) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub fn SWELL_osx_is_dark_mode(&self, mode: c_int) -> bool

impl Swell

This impl block contains functions which exist in SWELL as macros and therefore are not picked up by bindgen.

pub fn make_available_globally(functions: Swell)

Makes the given instance available globally.

After this has been called, the instance can be queried globally using get().

This can be called once only. Subsequent calls won’t have any effect!

pub fn get() -> &'static Swell

Gives access to the instance which you made available globally before.

Panics

This panics if make_available_globally() has not been called before.

pub fn pointers(&self) -> &SwellFunctionPointers

Gives access to the SWELL function pointers.

pub fn plugin_context(&self) -> &PluginContext

Returns the plug-in context.

pub unsafe fn CreateDialogParam(
&self,
hinst: HINSTANCE,
resid: *const c_char,
par: HWND,
dlgproc: Option<unsafe extern "C" fn(arg1: HWND, arg2: UINT, arg3: WPARAM, arg4: LPARAM) -> INT_PTR>,
param: LPARAM
) -> HWND

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn LoadMenu(&self, hinst: HINSTANCE, resid: *const c_char) -> HMENU

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn FillRect(&self, ctx: HDC, r: *const RECT, br: HBRUSH)

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn DrawText(
&self,
ctx: HDC,
buf: *const c_char,
len: c_int,
r: *mut RECT,
align: c_int
) -> c_int

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn SetWindowText(&self, hwnd: HWND, text: *const c_char) -> c_schar

Safety

REAPER can crash if you pass an invalid pointer.

pub unsafe fn GetWindowText(
&self,
hwnd: HWND,
lpString: LPSTR,
nMaxCount: c_int
) -> c_schar

Attention: Whereas the Windows original returns a length, this just returns success.

In order to avoid surprises, on Windows it will behave like this, too.

Safety

REAPER can crash if you pass an invalid pointer.

pub fn RGB(r: u8, g: u8, b: u8) -> c_uint

pub fn GetRValue(color: c_uint) -> u8

pub fn GetGValue(color: c_uint) -> u8

pub fn GetBValue(color: c_uint) -> u8