Trait reaper_low::ProjectStateContext

pub trait ProjectStateContext {
    fn AddLine(&mut self, line: *const c_char);
    fn GetLine(&mut self, buf: *mut c_char, buflen: c_int) -> c_int;
    fn GetOutputSize(&mut self) -> c_longlong;
    fn GetTempFlag(&mut self) -> c_int;
    fn SetTempFlag(&mut self, flag: c_int);
}

This is the Rust analog to the C++ virtual base class ProjectStateContext.

An implementation of this trait can be passed to create_cpp_to_rust_project_state_context().

Attention: Not really usable yet due to the lack of the variadic parameter in AddLine.

Required Methods

fn AddLine(&mut self, line: *const c_char)

fn GetLine(&mut self, buf: *mut c_char, buflen: c_int) -> c_int

fn GetOutputSize(&mut self) -> c_longlong

fn GetTempFlag(&mut self) -> c_int

fn SetTempFlag(&mut self, flag: c_int)